+++ client/src/PhotoUpload.jsx
@@ -1,27 +1,21 @@
-import React, { useState } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 
-const PhotoUpload = () => {
-  const [file, setFile] = useState(null);
-
-  const changeHandler = (e) => {
-    setFile(null);
-    e.persist();
-    if (e.target.files) {
-      const newImage = {
-        name: e.target.files[0].name,
-        url: URL.createObjectURL(e.target.files[0]),
-        file: e.target.files[0],
-      };
-      setFile(newImage);
-    }
-  };
-
-  return (
-    <div>
-      <input onChange={changeHandler} type="file" accept=".jpg, .png, .jpeg" />
-      {!file ? 'Select an Image' : <img src={file.url} alt="upload" />}
-    </div>
-  );
-};
+/**
+ * Selects a photo to upload from the user's computer
+ * @param {Function} changeHandler function that changes the file
+ * @param {File} file the file that is being uploaded 
+ */
+const PhotoUpload = ({ changeHandler, file }) => (
+  <div>
+    <input onChange={changeHandler} type="file" accept=".jpg, .png, .jpeg" />
+    {!file ? 'Select an Image' : <img src={file.url} alt="upload" />}
+  </div>
+);
 
 export default PhotoUpload;
+
+PhotoUpload.propTypes = {
+  changeHandler: PropTypes.func.isRequired,
+  file: PropTypes.array,
+};
